// client/src/music_transfer/song_list_item.ts
export interface SongListItem {
    title: string;
    artist: string;
    description: string;
    runtime: string;
    foundMatch: string;
    spotifySongTitle: string,
    spotifyArtists: string,
    spotifyAlbum: string,
    shouldUpload: string;
    savedToSpotify: string,
    spotifySongId: string,
    matchPercentage: string;
  }